+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - #427: Add `generateOptparseApplicativeCompletions`
     - #455: Add `drvAttrs` option for setting arbitrary derivation attributes
     - #459: Apply `custom` ahead of other settings.
+    - #469: Add `libraryHaskellDepends`, `librarySystemDepends`, `libraryToolDepends`, and `libraryPkgconfigDepends`
 
 ## 0.5.0 (Jun 24, 2024)